requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.8
Jinja2==2.7.3
SQLAlchemy==0.9.8
WTForms==2.0.2
Werkzeug==0.10.1
wsgiref==0.1.2
python-twitter==2.2
twitter-oauth==0.1.0
